Fixed backend
Adjusted backend for it to work with all database tables
